moved to an integration test
