+++ contracts/axelar-soroban-interfaces/Cargo.toml
@@ -11,5 +11,7 @@ axelar-soroban-std = { workspace = true }
 soroban-sdk = { workspace = true }
 
 [dev-dependencies]
+soroban-sdk = { workspace = true, features = ["testutils"] }
 goldie = { version = "0.5" }
-hex-literal = "0.3"
+hex-literal = "0.4"
+hex = "0.4"
